feat(dashboard): serve trace session artifacts
