+++ apps/dashboard/package.json
@@ -10,6 +10,7 @@
     "test": "bun test"
   },
   "dependencies": {
+    "@agentv/core": "workspace:*",
     "@monaco-editor/react": "^4.7.0",
     "@tanstack/react-query": "^5.75.5",
     "@tanstack/react-router": "^1.120.3",
