+++ packages/core/src/index.ts
@@ -2,6 +2,7 @@ export * from './evaluation/content.js';
 export * from './evaluation/types.js';
 export * from './evaluation/trace.js';
 export * from './evaluation/trace-envelope.js';
+export * from './evaluation/dashboard-trace-read-model.js';
 export * from './evaluation/external-trace.js';
 export * from './evaluation/projection-identity.js';
 export * from './evaluation/replay-fixtures.js';
